0622251126 Cosmeticly working

+++ main/control_led.h
@@ -7,7 +7,7 @@
 #define BLINK_GPIO CONFIG_BLINK_GPIO
 #define LED_BRIGHTNESS CONFIG_LED_BRIGHTNESS
 
-extern int led_state;
+extern uint8_t led_state;
 
 #ifdef CONFIG_BLINK_LED_STRIP
 void led_strip_hsv2rgb(uint32_t h, uint32_t s, uint32_t v, uint32_t *r, uint32_t *g, uint32_t *b);
